refactor: address PR review feedback on Phase 1
- Consolidate MaterialSupplier + PurchaseType into material_entity.go
  (domain and mysql layers) instead of separate files
- Change purchase_url to NOT NULL DEFAULT '' to avoid *string pointer
  typing and simplify transformer code
