+++ apps/api/data/mysql/material_entity.go
@@ -2,6 +2,17 @@ package mysql
 
 import "time"
 
+type MaterialSupplier struct {
+	Id           int64
+	MaterialId   int64
+	SupplierId   int64
+	Supplier     Supplier
+	PurchaseType string
+	PurchaseUrl  string
+	DeletedAt    *time.Time
+	CreatedAt    time.Time
+}
+
 type Material struct {
 	Id               int64
 	Name             string
