--- apps/api/data/mysql/material_supplier_entity.go
@@ -1,14 +1,0 @@
-package mysql
-
-import "time"
-
-type MaterialSupplier struct {
-	Id           int64
-	MaterialId   int64
-	SupplierId   int64
-	Supplier     Supplier
-	PurchaseType string
-	PurchaseUrl  *string
-	DeletedAt    *time.Time
-	CreatedAt    time.Time
-}
